close audioOutput every time playback is finished
The N900AudioOutput open()s the device for each call to start() or
resume().  However, it does not automatically close() the device
before.  As a result the pulseaudio handles run out after too many
calls to start().  Fix this by following the Qt example code and close
the audioOutput after every state change to a non-playing state.